LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
<p id="description"></p>
<div id="sandbox"></div>
<pre id="console"></pre>
<script>
description("Tests to ensure that event dispatching behaves as the Shadow DOM spec describes.");

var defaultPaddingSize = 40;

function moveMouseOver(element)
{
    if (!window.eventSender || !window.internals)
        return;

    var x = element.offsetLeft + element.offsetWidth / 2;
    var y;
    if (element.hasChildNodes() || window.internals.shadowRoot(element))
        y = element.offsetTop + defaultPaddingSize / 2;
    else
        y = element.offsetTop + element.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
}

var eventRecords = {};

function clearEventRecords()
{
    eventRecords = {};
}

function dispatchedEvent(eventType)
{
    var events = eventRecords[eventType];
    if (!events)
        return [];
    return events;
}

function recordEvent(event)
{
    var eventType = event.type
    if (!eventRecords[eventType]) {
        eventRecords[eventType] = []
    }
    var eventString = '';
    if (event.currentTarget)
        eventString += ' @' + event.currentTarget.id;
    if (event.target)
        eventString += ' (target: ' + event.target.id + ')';
    if (event.relatedTarget)
        eventString += ' (related: ' + event.relatedTarget.id + ')';
    if (event.eventPhase == 1)
        eventString += '(capturing phase)';
    if (event.target && event.currentTarget && event.target.id == event.currentTarget.id)
        shouldBe("event.eventPhase", "2", true);
    eventRecords[eventType].push(eventString);
}

function dumpNode(node)
{
    var output = node.nodeName + "\t";
    if (node.id)
        output += ' id=' + node.id;
    if (node.className)
        output += ' class=' + node.className;
    return output;
}

function dumpComposedShadowTree(node, indent)
{
    indent = indent || "";
    var output = indent + dumpNode(node) + "\n";
    var child;
    for (child = internals.firstChildByWalker(node); child; child = internals.nextSiblingByWalker(child))
         output += dumpComposedShadowTree(child, indent + "\t");
    return output;
}

function addEventListeners(nodes)
{
    for (var i = 0; i < nodes.length; ++i) {
        var node = getNodeInShadowTreeStack(nodes[i]);
        node.addEventListener('mouseover', recordEvent, false);
        node.addEventListener('mouseout', recordEvent, false);
        node.addEventListener('click', recordEvent, false);
        // <content> might be an inactive insertion point, so style it also.
        if (node.tagName == 'DIV' || node.tagName == 'DETAILS' || node.tagName == 'SUMMARY' || node.tagName == 'CONTENT')
            node.setAttribute('style', 'padding-top: ' + defaultPaddingSize + 'px;');
    }
}

function debugDispatchedEvent(eventType)
{
    debug('\n  ' + eventType);
    var events = dispatchedEvent(eventType);
    for (var i = 0; i < events.length; ++i)
        debug('    ' + events[i])
}

function moveMouse(oldElementId, newElementId)
{
    clearEventRecords();
    debug('\n' + 'Moving mouse from ' + oldElementId + ' to ' + newElementId);
    moveMouseOver(getNodeInShadowTreeStack(oldElementId));

    clearEventRecords();
    moveMouseOver(getNodeInShadowTreeStack(newElementId));

    debugDispatchedEvent('mouseout');
    debugDispatchedEvent('mouseover');
}

var sandbox = document.getElementById('sandbox');

function showSandboxTree()
{
    var sandbox = document.getElementById('sandbox');
    sandbox.offsetLeft;
    debug('\n\nComposed Shadow Tree will be:\n' + dumpComposedShadowTree(sandbox));
}

function testEventsOnDistributedChild()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'shadow-host'},
                            createShadowRoot(
                                createDOM('content', {'id': 'content', 'select': '#distributed-light-child',}),
                                createDOM('div', {'id': 'shadow-root-child'})),
                            createDOM('div', {'id': 'distributed-light-child'}),
                            createDOM('div', {'id': 'non-distributed-light-child'}))));

    addEventListeners(['top', 'shadow-host', 'shadow-host/', 'shadow-host/content', 'shadow-host/shadow-root-child',
                       'distributed-light-child', 'non-distributed-light-child']);
    getNodeInShadowTreeStack('shadow-host/').id = 'shadow-root';
    showSandboxTree();

    moveMouse('shadow-host', 'shadow-host/shadow-root-child');
    moveMouse('shadow-host/shadow-root-child', 'shadow-host');

    moveMouse('shadow-host', 'distributed-light-child');
    moveMouse('distributed-light-child', 'shadow-host');

    moveMouse('shadow-host/shadow-root-child', 'distributed-light-child');
    moveMouse('distributed-light-child', 'shadow-host/shadow-root-child');
}

function testEventsOnDetailsSummary()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  // 'details/summary' elements use Shadow DOM in its implementation.
                  createDOM('details', {'id': 'details', 'open': true},
                            createDOM('summary', {'id': 'summary'}))));

    addEventListeners(['top', 'details', 'summary']);
    showSandboxTree();

    moveMouse('details', 'summary');
    moveMouse('summary', 'details');
}

function testEventsOnNestedShadowRoots()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'A'},
                            createDOM('div', {'id': 'B'},
                                      createShadowRoot(
                                          createDOM('div', {'id': 'G'},
                                                    createShadowRoot(
                                                        createDOM('div', {'id': 'J'},
                                                                  createShadowRoot(
                                                                      createDOM('content', {'id': 'N', 'select': '#C'})),
                                                                  createDOM('content', {'id': 'K', 'select': '#C'})),
                                                        createDOM('div', {'id': 'L'},
                                                                  createShadowRoot(
                                                                      createDOM('content', {'id': 'O', 'select': '#E'})),
                                                                  createDOM('content', {'id': 'M', 'select': '#E'}))),
                                                    createDOM('content', {'id': 'H', 'select': '#C'}),
                                                    createDOM('content', {'id': 'I', 'select': '#E'}))),
                                      createDOM('div', {'id': 'C'},
                                                createDOM('div', {'id': 'D'})),
                                      createDOM('div', {'id': 'E'},
                                                createDOM('div', {'id': 'F'}))))));

    addEventListeners(['top', 'A', 'B', 'C', 'D', 'E', 'F', 'B/', 'B/G', 'B/H', 'B/I', 'B/G/', 'B/G/J', 'B/G/K', 'B/G/L', 'B/G/M',
                       'B/G/J/', 'B/G/J/N', 'B/G/L/', 'B/G/L/O']);
    getNodeInShadowTreeStack('B/').id = 'shadow-root-B';
    getNodeInShadowTreeStack('B/G/').id = 'shadow-root-G';
    getNodeInShadowTreeStack('B/G/J/').id = 'shadow-root-J';
    getNodeInShadowTreeStack('B/G/L/').id = 'shadow-root-L';
    showSandboxTree();

    moveMouse('F', 'D');
    moveMouse('B/G/L', 'D');
    moveMouse('B/G/L', 'B/G/J');
    moveMouse('A', 'D');
    moveMouse('D', 'A');
}

function testEventsOnSVGInShadowSubtree()
{
    // Makes sure that <svg> in shadow DOM subtree, which is not supported at this time, does not crash.
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'shadow-host'},
                            createShadowRoot())));
    var svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
    svg.id = 'svg-in-shadow-tree';
    getNodeInShadowTreeStack('shadow-host/').appendChild(svg);

    addEventListeners(['top', 'shadow-host/', 'shadow-host/svg-in-shadow-tree']);
    getNodeInShadowTreeStack('shadow-host/').id = 'shadow-root';
    showSandboxTree();

    moveMouse('shadow-host/svg-in-shadow-tree', 'top');
}

function testEventsOnTextNodeOfShadowRoot()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'shadow-host'},
                            createShadowRoot())));
    var shadowRoot = getNodeInShadowTreeStack('shadow-host/');
    shadowRoot.id = 'shadow-root';
    shadowRoot.innerHTML = 'Text Nodes';
    addEventListeners(['top', 'shadow-host', 'shadow-host/']);
    showSandboxTree();

    // Calculates the position of the text node in the shadow root.
    var host = document.getElementById('shadow-host');
    var x = host.offsetLeft + 5;
    var y = host.offsetTop + defaultPaddingSize + 5;
    debug('\n' + 'Moving mouse from a direct child text node of the shadow root to top');
    eventSender.mouseMoveTo(x, y);
    clearEventRecords();
    moveMouseOver(document.getElementById('top'));
    debugDispatchedEvent('mouseout');
    debugDispatchedEvent('mouseover');
}

function testEventsOnDistributedTextNode()
{
    // Makes sure an insertion point can receive a event when a distributed text node is clicked.
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'shadow-host'},
                            createShadowRoot(
                                createDOM('content', {'id': 'content'})),
                            document.createTextNode('Text Node'))));
    var shadowRoot = getNodeInShadowTreeStack('shadow-host/').id = 'shadow-root';
    addEventListeners(['top', 'shadow-host', 'shadow-host/', 'shadow-host/content']);
    showSandboxTree();

    // Calculates the position of the text node.
    var host = document.getElementById('shadow-host');
    var x = host.offsetLeft + 5;
    var y = host.offsetTop + defaultPaddingSize + 5;
    debug('\n' + 'Moving mouse from a distributed text node to top');
    eventSender.mouseMoveTo(x, y);
    clearEventRecords();
    moveMouseOver(document.getElementById('top'));
    debugDispatchedEvent('mouseout');
    debugDispatchedEvent('mouseover');
}

function testEventsOnChildOfInactiveContent()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'A'}),
                  createDOM('div', {'id': 'B'},
                            createShadowRoot(
                                createDOM('content', {'id': 'active-content', 'select': '#parent-of-inactive-content'})),
                            createDOM('div', {'id': 'parent-of-inactive-content'},
                                      createDOM('content', {'id': 'inactive-content'},
                                                createDOM('div', {'id': 'child-of-inactive-content'}))))));

    addEventListeners(['top', 'A', 'B', 'B/', 'B/active-content',
                       'parent-of-inactive-content', 'inactive-content', 'child-of-inactive-content']);
    getNodeInShadowTreeStack('B/').id = 'shadow-root-B';
    showSandboxTree();
    moveMouse('A', 'child-of-inactive-content');
}

function testEventsOnMultipleShadowRoots()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'A'},
                            createShadowRoot(
                                createDOM('content', {'id': 'C'}),
                                createDOM('div', {'id': 'D'})),
                            createShadowRoot(
                                createDOM('shadow', {'id': 'E'}),
                                createDOM('div', {'id': 'F'})),
                            createDOM('div', {'id': 'B'}))));

    addEventListeners(['top', 'A', 'B', 'A/', 'A/C', 'A/D', 'A//', 'A//E', 'A//F']);
    getNodeInShadowTreeStack('A/').id = 'older-shadow-root';
    getNodeInShadowTreeStack('A//').id = 'younger-shadow-root';
    showSandboxTree();

    moveMouse('B', 'A');
    moveMouse('A/D', 'A//F');
    moveMouse('B', 'A//F');
}

function testEventsOnNonDistributedNodes()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'A'},
                            createShadowRoot(
                                createDOM('div', {'id': 'B'},
                                          createDOM('div', {'id': 'C'}))),
                            createShadowRoot(
                                createDOM('div', {'id': 'D'})),
                            createDOM('div', {'id': 'non-distributed-node'}))));

    addEventListeners(['top', 'A', 'A/', 'A/B', 'A/C', 'A//', 'A//D', 'non-distributed-node']);
    getNodeInShadowTreeStack('A/').id = 'orphaned-shadow-root';
    getNodeInShadowTreeStack('A//').id = 'youngest-shadow-root';
    parent.offsetLeft;
    showSandboxTree();

    clearEventRecords();
    debug('\nClick C');
    getNodeInShadowTreeStack('A/C').click();
    debugDispatchedEvent('click');

    clearEventRecords();
    debug('\nClick non-distributed-node');
    document.getElementById('non-distributed-node').click();
    debugDispatchedEvent('click');
}

function testEventsOnFallbackElements()
{
    sandbox.innerHTML = '';
    sandbox.appendChild(
        createDOM('div', {'id': 'top'},
                  createDOM('div', {'id': 'A'},
                            createShadowRoot(
                                createDOM('content', {'id': 'content1', 'select': '#none'},
                                          createDOM('div', {'id': 'used-fallback'})),
                                createDOM('content', {'id': 'content2'},
                                          createDOM('div', {'id': 'non-used-fallback'}))),
                            createDOM('div', {'id': 'B'}))));

    addEventListeners(['top', 'A', 'B', 'A/', 'A/content1', 'A/used-fallback', 'A/content2', 'A/non-used-fallback']);
    getNodeInShadowTreeStack('A/').id = 'shadow-root';
    showSandboxTree();

    moveMouse('A/used-fallback', 'A');
    moveMouse('A', 'A/used-fallback');

    clearEventRecords();
    debug('\nClick non-used-fallback node');
    getNodeInShadowTreeStack('A/non-used-fallback').click();
    debugDispatchedEvent('click');
}

function test()
{
    if (window.testRunner)
        testRunner.dumpAsText();

    testEventsOnDistributedChild();
    testEventsOnDetailsSummary();
    testEventsOnNestedShadowRoots();
    testEventsOnSVGInShadowSubtree();
    testEventsOnTextNodeOfShadowRoot();
    testEventsOnDistributedTextNode();
    testEventsOnChildOfInactiveContent();
    testEventsOnMultipleShadowRoots();
    testEventsOnNonDistributedNodes();
    testEventsOnFallbackElements();
}

test();
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
